+++ CHANGELOG.md
@@ -21,6 +21,7 @@ FlatLaf Change Log
   `StringBuilder`, `StringBuffer`, or any other object that returns HTML text in
   method `toString()`. (similar to issue #12)
 - ComboBox: Fixed popup border painting on HiDPI screens (e.g. at 150% scaling).
+- ComboBox: Fixed popup location if shown above of combo box (Java 8 only).
 - ComboBox (editable): Fixed wrong border of internal text field under special
   circumstances.
 - Spinner: Fixed painting of border corners on left side. (issue #382;